+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 - AlfreedPage get alfreedpageBuilder is now build()
 - get viewmodel from any alfreedPage child
 - fix lint errors
-- 
+- optmized alfreed page builder 
 
 ## 0.3.0
 - Remove unnecessary null check